Use getOrSetTypeResolver to link python with direct c++ generated Alternatives
