+++ typed_python/AllTypes.hpp
@@ -46,3 +46,4 @@
 
 #include "direct_types/all.hpp"
 #include "direct_types/GeneratedTypes.hpp"
+#include "direct_types/ClientToServer.hpp"
