+++ typed_python/__init__.py
@@ -22,7 +22,8 @@ from typed_python.type_filter import TypeFilter
 from typed_python._types import (
     TupleOf, ListOf, Tuple, NamedTuple, OneOf, ConstDict,
     Alternative, Value, serialize, deserialize, serializeStream, deserializeStream,
-    PointerTo, Dict, validateSerializedObject, validateSerializedObjectStream, decodeSerializedObject
+    PointerTo, Dict, validateSerializedObject, validateSerializedObjectStream, decodeSerializedObject,
+    getOrSetTypeResolver
 )
 import typed_python._types as _types
