+++ typed_python/direct_types/GeneratedTypes.hpp
@@ -1,109 +1,8 @@
 #pragma once
-// Generated Tuple Anon39611216
-//    a0=String
-//    a1=Bytes
-class Anon39611216 {
-public:
-    typedef String a0_type;
-    typedef Bytes a1_type;
-    a0_type& a0() const { return *(a0_type*)(data); }
-    a1_type& a1() const { return *(a1_type*)(data + size1); }
-    static Tuple* getType() {
-        static Tuple* t = Tuple::Make({
-                TypeDetails<Anon39611216::a0_type>::getType(),
-                TypeDetails<Anon39611216::a1_type>::getType()
-            });
-        return t;
-        }
-
-    static Anon39611216 fromPython(PyObject* p) {
-        Anon39611216 l;
-        PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
-        return l;
-    }
-
-    PyObject* toPython() {
-        return PyInstance::extractPythonObject((instance_ptr)this, getType());
-    }
-
-    Anon39611216& operator = (const Anon39611216& other) {
-        a0() = other.a0();
-        a1() = other.a1();
-        return *this;
-    }
-
-    Anon39611216(const Anon39611216& other) {
-        new (&a0()) a0_type(other.a0());
-        new (&a1()) a1_type(other.a1());
-    }
-
-    ~Anon39611216() {
-        a1().~a1_type();
-        a0().~a0_type();
-    }
-
-    Anon39611216() {
-        bool inita0 = false;
-        bool inita1 = false;
-        try {
-            new (&a0()) a0_type();
-            inita0 = true;
-            new (&a1()) a1_type();
-            inita1 = true;
-        } catch(...) {
-            try {
-                if (inita1) a1().~a1_type();
-                if (inita0) a0().~a0_type();
-            } catch(...) {
-            }
-            throw;
-        }
-    }
-
-    Anon39611216(const a0_type& a0_val, const a1_type& a1_val) {
-        bool inita0 = false;
-        bool inita1 = false;
-        try {
-            new (&a0()) a0_type(a0_val);
-            inita0 = true;
-            new (&a1()) a1_type(a1_val);
-            inita1 = true;
-        } catch(...) {
-            try {
-                if (inita1) a1().~a1_type();
-                if (inita0) a0().~a0_type();
-            } catch(...) {
-            }
-            throw;
-        }
-    }
-private:
-    static const int size1 = sizeof(a0_type);
-    static const int size2 = sizeof(a1_type);
-    uint8_t data[size1 + size2];
-};
-
-template <>
-class TypeDetails<Anon39611216> {
-public:
-    static Type* getType() {
-        static Type* t = Anon39611216::getType();
-        if (t->bytecount() != bytecount) {
-            throw std::runtime_error("Anon39611216 somehow we have the wrong bytecount!");
-        }
-        return t;
-    }
-    static const uint64_t bytecount = 
-        sizeof(Anon39611216::a0_type) +
-        sizeof(Anon39611216::a1_type);
-};
-
-// END Generated Tuple Anon39611216
-
-// Generated Tuple Anon41033936
+// Generated Tuple Anon35575456
 //    a0=int64_t
 //    a1=int64_t
-class Anon41033936 {
+class Anon35575456 {
 public:
     typedef int64_t a0_type;
     typedef int64_t a1_type;
@@ -111,14 +10,14 @@ public:
     a1_type& a1() const { return *(a1_type*)(data + size1); }
     static Tuple* getType() {
         static Tuple* t = Tuple::Make({
-                TypeDetails<Anon41033936::a0_type>::getType(),
-                TypeDetails<Anon41033936::a1_type>::getType()
+                TypeDetails<Anon35575456::a0_type>::getType(),
+                TypeDetails<Anon35575456::a1_type>::getType()
             });
         return t;
         }
 
-    static Anon41033936 fromPython(PyObject* p) {
-        Anon41033936 l;
+    static Anon35575456 fromPython(PyObject* p) {
+        Anon35575456 l;
         PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
         return l;
     }
@@ -127,23 +26,23 @@ public:
         return PyInstance::extractPythonObject((instance_ptr)this, getType());
     }
 
-    Anon41033936& operator = (const Anon41033936& other) {
+    Anon35575456& operator = (const Anon35575456& other) {
         a0() = other.a0();
         a1() = other.a1();
         return *this;
     }
 
-    Anon41033936(const Anon41033936& other) {
+    Anon35575456(const Anon35575456& other) {
         new (&a0()) a0_type(other.a0());
         new (&a1()) a1_type(other.a1());
     }
 
-    ~Anon41033936() {
+    ~Anon35575456() {
         a1().~a1_type();
         a0().~a0_type();
     }
 
-    Anon41033936() {
+    Anon35575456() {
         bool inita0 = false;
         bool inita1 = false;
         try {
@@ -161,7 +60,7 @@ public:
         }
     }
 
-    Anon41033936(const a0_type& a0_val, const a1_type& a1_val) {
+    Anon35575456(const a0_type& a0_val, const a1_type& a1_val) {
         bool inita0 = false;
         bool inita1 = false;
         try {
@@ -185,26 +84,26 @@ private:
 };
 
 template <>
-class TypeDetails<Anon41033936> {
+class TypeDetails<Anon35575456> {
 public:
     static Type* getType() {
-        static Type* t = Anon41033936::getType();
+        static Type* t = Anon35575456::getType();
         if (t->bytecount() != bytecount) {
-            throw std::runtime_error("Anon41033936 somehow we have the wrong bytecount!");
+            throw std::runtime_error("Anon35575456 somehow we have the wrong bytecount!");
         }
         return t;
     }
     static const uint64_t bytecount = 
-        sizeof(Anon41033936::a0_type) +
-        sizeof(Anon41033936::a1_type);
+        sizeof(Anon35575456::a0_type) +
+        sizeof(Anon35575456::a1_type);
 };
 
-// END Generated Tuple Anon41033936
+// END Generated Tuple Anon35575456
 
-// Generated Tuple Anon41042080
+// Generated Tuple Anon35583280
 //    a0=bool
 //    a1=bool
-class Anon41042080 {
+class Anon35583280 {
 public:
     typedef bool a0_type;
     typedef bool a1_type;
@@ -212,14 +111,14 @@ public:
     a1_type& a1() const { return *(a1_type*)(data + size1); }
     static Tuple* getType() {
         static Tuple* t = Tuple::Make({
-                TypeDetails<Anon41042080::a0_type>::getType(),
-                TypeDetails<Anon41042080::a1_type>::getType()
+                TypeDetails<Anon35583280::a0_type>::getType(),
+                TypeDetails<Anon35583280::a1_type>::getType()
             });
         return t;
         }
 
-    static Anon41042080 fromPython(PyObject* p) {
-        Anon41042080 l;
+    static Anon35583280 fromPython(PyObject* p) {
+        Anon35583280 l;
         PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
         return l;
     }
@@ -228,23 +127,23 @@ public:
         return PyInstance::extractPythonObject((instance_ptr)this, getType());
     }
 
-    Anon41042080& operator = (const Anon41042080& other) {
+    Anon35583280& operator = (const Anon35583280& other) {
         a0() = other.a0();
         a1() = other.a1();
         return *this;
     }
 
-    Anon41042080(const Anon41042080& other) {
+    Anon35583280(const Anon35583280& other) {
         new (&a0()) a0_type(other.a0());
         new (&a1()) a1_type(other.a1());
     }
 
-    ~Anon41042080() {
+    ~Anon35583280() {
         a1().~a1_type();
         a0().~a0_type();
     }
 
-    Anon41042080() {
+    Anon35583280() {
         bool inita0 = false;
         bool inita1 = false;
         try {
@@ -262,7 +161,7 @@ public:
         }
     }
 
-    Anon41042080(const a0_type& a0_val, const a1_type& a1_val) {
+    Anon35583280(const a0_type& a0_val, const a1_type& a1_val) {
         bool inita0 = false;
         bool inita1 = false;
         try {
@@ -286,26 +185,26 @@ private:
 };
 
 template <>
-class TypeDetails<Anon41042080> {
+class TypeDetails<Anon35583280> {
 public:
     static Type* getType() {
-        static Type* t = Anon41042080::getType();
+        static Type* t = Anon35583280::getType();
         if (t->bytecount() != bytecount) {
-            throw std::runtime_error("Anon41042080 somehow we have the wrong bytecount!");
+            throw std::runtime_error("Anon35583280 somehow we have the wrong bytecount!");
         }
         return t;
     }
     static const uint64_t bytecount = 
-        sizeof(Anon41042080::a0_type) +
-        sizeof(Anon41042080::a1_type);
+        sizeof(Anon35583280::a0_type) +
+        sizeof(Anon35583280::a1_type);
 };
 
-// END Generated Tuple Anon41042080
+// END Generated Tuple Anon35583280
 
-// Generated NamedTuple Anon41057728
+// Generated NamedTuple Anon35597888
 //    x=int64_t
 //    y=int64_t
-class Anon41057728 {
+class Anon35597888 {
 public:
     typedef int64_t x_type;
     typedef int64_t y_type;
@@ -313,8 +212,8 @@ public:
     y_type& y() const { return *(y_type*)(data + size1); }
     static NamedTuple* getType() {
         static NamedTuple* t = NamedTuple::Make({
-                TypeDetails<Anon41057728::x_type>::getType(),
-                TypeDetails<Anon41057728::y_type>::getType()
+                TypeDetails<Anon35597888::x_type>::getType(),
+                TypeDetails<Anon35597888::y_type>::getType()
             },{
                 "x",
                 "y"
@@ -322,8 +221,8 @@ public:
         return t;
         }
 
-    static Anon41057728 fromPython(PyObject* p) {
-        Anon41057728 l;
+    static Anon35597888 fromPython(PyObject* p) {
+        Anon35597888 l;
         PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
         return l;
     }
@@ -332,23 +231,23 @@ public:
         return PyInstance::extractPythonObject((instance_ptr)this, getType());
     }
 
-    Anon41057728& operator = (const Anon41057728& other) {
+    Anon35597888& operator = (const Anon35597888& other) {
         x() = other.x();
         y() = other.y();
         return *this;
     }
 
-    Anon41057728(const Anon41057728& other) {
+    Anon35597888(const Anon35597888& other) {
         new (&x()) x_type(other.x());
         new (&y()) y_type(other.y());
     }
 
-    ~Anon41057728() {
+    ~Anon35597888() {
         y().~y_type();
         x().~x_type();
     }
 
-    Anon41057728() {
+    Anon35597888() {
         bool initx = false;
         bool inity = false;
         try {
@@ -366,7 +265,7 @@ public:
         }
     }
 
-    Anon41057728(const x_type& x_val, const y_type& y_val) {
+    Anon35597888(const x_type& x_val, const y_type& y_val) {
         bool initx = false;
         bool inity = false;
         try {
@@ -390,1031 +289,21 @@ private:
 };
 
 template <>
-class TypeDetails<Anon41057728> {
+class TypeDetails<Anon35597888> {
 public:
     static Type* getType() {
-        static Type* t = Anon41057728::getType();
+        static Type* t = Anon35597888::getType();
         if (t->bytecount() != bytecount) {
-            throw std::runtime_error("Anon41057728 somehow we have the wrong bytecount!");
+            throw std::runtime_error("Anon35597888 somehow we have the wrong bytecount!");
         }
         return t;
     }
     static const uint64_t bytecount = 
-        sizeof(Anon41057728::x_type) +
-        sizeof(Anon41057728::y_type);
+        sizeof(Anon35597888::x_type) +
+        sizeof(Anon35597888::y_type);
 };
 
-// END Generated NamedTuple Anon41057728
-
-// Generated NamedTuple ObjectFieldId
-//    objId=int64_t
-//    fieldId=int64_t
-//    isIndexValue=bool
-class ObjectFieldId {
-public:
-    typedef int64_t objId_type;
-    typedef int64_t fieldId_type;
-    typedef bool isIndexValue_type;
-    objId_type& objId() const { return *(objId_type*)(data); }
-    fieldId_type& fieldId() const { return *(fieldId_type*)(data + size1); }
-    isIndexValue_type& isIndexValue() const { return *(isIndexValue_type*)(data + size1 + size2); }
-    static NamedTuple* getType() {
-        static NamedTuple* t = NamedTuple::Make({
-                TypeDetails<ObjectFieldId::objId_type>::getType(),
-                TypeDetails<ObjectFieldId::fieldId_type>::getType(),
-                TypeDetails<ObjectFieldId::isIndexValue_type>::getType()
-            },{
-                "objId",
-                "fieldId",
-                "isIndexValue"
-            });
-        return t;
-        }
-
-    static ObjectFieldId fromPython(PyObject* p) {
-        ObjectFieldId l;
-        PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
-        return l;
-    }
-
-    PyObject* toPython() {
-        return PyInstance::extractPythonObject((instance_ptr)this, getType());
-    }
-
-    ObjectFieldId& operator = (const ObjectFieldId& other) {
-        objId() = other.objId();
-        fieldId() = other.fieldId();
-        isIndexValue() = other.isIndexValue();
-        return *this;
-    }
-
-    ObjectFieldId(const ObjectFieldId& other) {
-        new (&objId()) objId_type(other.objId());
-        new (&fieldId()) fieldId_type(other.fieldId());
-        new (&isIndexValue()) isIndexValue_type(other.isIndexValue());
-    }
-
-    ~ObjectFieldId() {
-        isIndexValue().~isIndexValue_type();
-        fieldId().~fieldId_type();
-        objId().~objId_type();
-    }
-
-    ObjectFieldId() {
-        bool initobjId = false;
-        bool initfieldId = false;
-        bool initisIndexValue = false;
-        try {
-            new (&objId()) objId_type();
-            initobjId = true;
-            new (&fieldId()) fieldId_type();
-            initfieldId = true;
-            new (&isIndexValue()) isIndexValue_type();
-            initisIndexValue = true;
-        } catch(...) {
-            try {
-                if (initisIndexValue) isIndexValue().~isIndexValue_type();
-                if (initfieldId) fieldId().~fieldId_type();
-                if (initobjId) objId().~objId_type();
-            } catch(...) {
-            }
-            throw;
-        }
-    }
-
-    ObjectFieldId(const objId_type& objId_val, const fieldId_type& fieldId_val, const isIndexValue_type& isIndexValue_val) {
-        bool initobjId = false;
-        bool initfieldId = false;
-        bool initisIndexValue = false;
-        try {
-            new (&objId()) objId_type(objId_val);
-            initobjId = true;
-            new (&fieldId()) fieldId_type(fieldId_val);
-            initfieldId = true;
-            new (&isIndexValue()) isIndexValue_type(isIndexValue_val);
-            initisIndexValue = true;
-        } catch(...) {
-            try {
-                if (initisIndexValue) isIndexValue().~isIndexValue_type();
-                if (initfieldId) fieldId().~fieldId_type();
-                if (initobjId) objId().~objId_type();
-            } catch(...) {
-            }
-            throw;
-        }
-    }
-private:
-    static const int size1 = sizeof(objId_type);
-    static const int size2 = sizeof(fieldId_type);
-    static const int size3 = sizeof(isIndexValue_type);
-    uint8_t data[size1 + size2 + size3];
-};
-
-template <>
-class TypeDetails<ObjectFieldId> {
-public:
-    static Type* getType() {
-        static Type* t = ObjectFieldId::getType();
-        if (t->bytecount() != bytecount) {
-            throw std::runtime_error("ObjectFieldId somehow we have the wrong bytecount!");
-        }
-        return t;
-    }
-    static const uint64_t bytecount = 
-        sizeof(ObjectFieldId::objId_type) +
-        sizeof(ObjectFieldId::fieldId_type) +
-        sizeof(ObjectFieldId::isIndexValue_type);
-};
-
-// END Generated NamedTuple ObjectFieldId
-
-// Generated NamedTuple IndexId
-//    fieldId=int64_t
-//    indexValue=Bytes
-class IndexId {
-public:
-    typedef int64_t fieldId_type;
-    typedef Bytes indexValue_type;
-    fieldId_type& fieldId() const { return *(fieldId_type*)(data); }
-    indexValue_type& indexValue() const { return *(indexValue_type*)(data + size1); }
-    static NamedTuple* getType() {
-        static NamedTuple* t = NamedTuple::Make({
-                TypeDetails<IndexId::fieldId_type>::getType(),
-                TypeDetails<IndexId::indexValue_type>::getType()
-            },{
-                "fieldId",
-                "indexValue"
-            });
-        return t;
-        }
-
-    static IndexId fromPython(PyObject* p) {
-        IndexId l;
-        PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
-        return l;
-    }
-
-    PyObject* toPython() {
-        return PyInstance::extractPythonObject((instance_ptr)this, getType());
-    }
-
-    IndexId& operator = (const IndexId& other) {
-        fieldId() = other.fieldId();
-        indexValue() = other.indexValue();
-        return *this;
-    }
-
-    IndexId(const IndexId& other) {
-        new (&fieldId()) fieldId_type(other.fieldId());
-        new (&indexValue()) indexValue_type(other.indexValue());
-    }
-
-    ~IndexId() {
-        indexValue().~indexValue_type();
-        fieldId().~fieldId_type();
-    }
-
-    IndexId() {
-        bool initfieldId = false;
-        bool initindexValue = false;
-        try {
-            new (&fieldId()) fieldId_type();
-            initfieldId = true;
-            new (&indexValue()) indexValue_type();
-            initindexValue = true;
-        } catch(...) {
-            try {
-                if (initindexValue) indexValue().~indexValue_type();
-                if (initfieldId) fieldId().~fieldId_type();
-            } catch(...) {
-            }
-            throw;
-        }
-    }
-
-    IndexId(const fieldId_type& fieldId_val, const indexValue_type& indexValue_val) {
-        bool initfieldId = false;
-        bool initindexValue = false;
-        try {
-            new (&fieldId()) fieldId_type(fieldId_val);
-            initfieldId = true;
-            new (&indexValue()) indexValue_type(indexValue_val);
-            initindexValue = true;
-        } catch(...) {
-            try {
-                if (initindexValue) indexValue().~indexValue_type();
-                if (initfieldId) fieldId().~fieldId_type();
-            } catch(...) {
-            }
-            throw;
-        }
-    }
-private:
-    static const int size1 = sizeof(fieldId_type);
-    static const int size2 = sizeof(indexValue_type);
-    uint8_t data[size1 + size2];
-};
-
-template <>
-class TypeDetails<IndexId> {
-public:
-    static Type* getType() {
-        static Type* t = IndexId::getType();
-        if (t->bytecount() != bytecount) {
-            throw std::runtime_error("IndexId somehow we have the wrong bytecount!");
-        }
-        return t;
-    }
-    static const uint64_t bytecount = 
-        sizeof(IndexId::fieldId_type) +
-        sizeof(IndexId::indexValue_type);
-};
-
-// END Generated NamedTuple IndexId
-
-// Generated NamedTuple FieldDefinition
-//    schema=String
-//    typename0=String
-//    fieldname=String
-class FieldDefinition {
-public:
-    typedef String schema_type;
-    typedef String typename0_type;
-    typedef String fieldname_type;
-    schema_type& schema() const { return *(schema_type*)(data); }
-    typename0_type& typename0() const { return *(typename0_type*)(data + size1); }
-    fieldname_type& fieldname() const { return *(fieldname_type*)(data + size1 + size2); }
-    static NamedTuple* getType() {
-        static NamedTuple* t = NamedTuple::Make({
-                TypeDetails<FieldDefinition::schema_type>::getType(),
-                TypeDetails<FieldDefinition::typename0_type>::getType(),
-                TypeDetails<FieldDefinition::fieldname_type>::getType()
-            },{
-                "schema",
-                "typename0",
-                "fieldname"
-            });
-        return t;
-        }
-
-    static FieldDefinition fromPython(PyObject* p) {
-        FieldDefinition l;
-        PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
-        return l;
-    }
-
-    PyObject* toPython() {
-        return PyInstance::extractPythonObject((instance_ptr)this, getType());
-    }
-
-    FieldDefinition& operator = (const FieldDefinition& other) {
-        schema() = other.schema();
-        typename0() = other.typename0();
-        fieldname() = other.fieldname();
-        return *this;
-    }
-
-    FieldDefinition(const FieldDefinition& other) {
-        new (&schema()) schema_type(other.schema());
-        new (&typename0()) typename0_type(other.typename0());
-        new (&fieldname()) fieldname_type(other.fieldname());
-    }
-
-    ~FieldDefinition() {
-        fieldname().~fieldname_type();
-        typename0().~typename0_type();
-        schema().~schema_type();
-    }
-
-    FieldDefinition() {
-        bool initschema = false;
-        bool inittypename0 = false;
-        bool initfieldname = false;
-        try {
-            new (&schema()) schema_type();
-            initschema = true;
-            new (&typename0()) typename0_type();
-            inittypename0 = true;
-            new (&fieldname()) fieldname_type();
-            initfieldname = true;
-        } catch(...) {
-            try {
-                if (initfieldname) fieldname().~fieldname_type();
-                if (inittypename0) typename0().~typename0_type();
-                if (initschema) schema().~schema_type();
-            } catch(...) {
-            }
-            throw;
-        }
-    }
-
-    FieldDefinition(const schema_type& schema_val, const typename0_type& typename0_val, const fieldname_type& fieldname_val) {
-        bool initschema = false;
-        bool inittypename0 = false;
-        bool initfieldname = false;
-        try {
-            new (&schema()) schema_type(schema_val);
-            initschema = true;
-            new (&typename0()) typename0_type(typename0_val);
-            inittypename0 = true;
-            new (&fieldname()) fieldname_type(fieldname_val);
-            initfieldname = true;
-        } catch(...) {
-            try {
-                if (initfieldname) fieldname().~fieldname_type();
-                if (inittypename0) typename0().~typename0_type();
-                if (initschema) schema().~schema_type();
-            } catch(...) {
-            }
-            throw;
-        }
-    }
-private:
-    static const int size1 = sizeof(schema_type);
-    static const int size2 = sizeof(typename0_type);
-    static const int size3 = sizeof(fieldname_type);
-    uint8_t data[size1 + size2 + size3];
-};
-
-template <>
-class TypeDetails<FieldDefinition> {
-public:
-    static Type* getType() {
-        static Type* t = FieldDefinition::getType();
-        if (t->bytecount() != bytecount) {
-            throw std::runtime_error("FieldDefinition somehow we have the wrong bytecount!");
-        }
-        return t;
-    }
-    static const uint64_t bytecount = 
-        sizeof(FieldDefinition::schema_type) +
-        sizeof(FieldDefinition::typename0_type) +
-        sizeof(FieldDefinition::fieldname_type);
-};
-
-// END Generated NamedTuple FieldDefinition
-
-// Generated NamedTuple TypeDefinition
-//    fields=TupleOf<String>
-//    indices=TupleOf<String>
-class TypeDefinition {
-public:
-    typedef TupleOf<String> fields_type;
-    typedef TupleOf<String> indices_type;
-    fields_type& fields() const { return *(fields_type*)(data); }
-    indices_type& indices() const { return *(indices_type*)(data + size1); }
-    static NamedTuple* getType() {
-        static NamedTuple* t = NamedTuple::Make({
-                TypeDetails<TypeDefinition::fields_type>::getType(),
-                TypeDetails<TypeDefinition::indices_type>::getType()
-            },{
-                "fields",
-                "indices"
-            });
-        return t;
-        }
-
-    static TypeDefinition fromPython(PyObject* p) {
-        TypeDefinition l;
-        PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
-        return l;
-    }
-
-    PyObject* toPython() {
-        return PyInstance::extractPythonObject((instance_ptr)this, getType());
-    }
-
-    TypeDefinition& operator = (const TypeDefinition& other) {
-        fields() = other.fields();
-        indices() = other.indices();
-        return *this;
-    }
-
-    TypeDefinition(const TypeDefinition& other) {
-        new (&fields()) fields_type(other.fields());
-        new (&indices()) indices_type(other.indices());
-    }
-
-    ~TypeDefinition() {
-        indices().~indices_type();
-        fields().~fields_type();
-    }
-
-    TypeDefinition() {
-        bool initfields = false;
-        bool initindices = false;
-        try {
-            new (&fields()) fields_type();
-            initfields = true;
-            new (&indices()) indices_type();
-            initindices = true;
-        } catch(...) {
-            try {
-                if (initindices) indices().~indices_type();
-                if (initfields) fields().~fields_type();
-            } catch(...) {
-            }
-            throw;
-        }
-    }
-
-    TypeDefinition(const fields_type& fields_val, const indices_type& indices_val) {
-        bool initfields = false;
-        bool initindices = false;
-        try {
-            new (&fields()) fields_type(fields_val);
-            initfields = true;
-            new (&indices()) indices_type(indices_val);
-            initindices = true;
-        } catch(...) {
-            try {
-                if (initindices) indices().~indices_type();
-                if (initfields) fields().~fields_type();
-            } catch(...) {
-            }
-            throw;
-        }
-    }
-private:
-    static const int size1 = sizeof(fields_type);
-    static const int size2 = sizeof(indices_type);
-    uint8_t data[size1 + size2];
-};
-
-template <>
-class TypeDetails<TypeDefinition> {
-public:
-    static Type* getType() {
-        static Type* t = TypeDefinition::getType();
-        if (t->bytecount() != bytecount) {
-            throw std::runtime_error("TypeDefinition somehow we have the wrong bytecount!");
-        }
-        return t;
-    }
-    static const uint64_t bytecount = 
-        sizeof(TypeDefinition::fields_type) +
-        sizeof(TypeDefinition::indices_type);
-};
-
-// END Generated NamedTuple TypeDefinition
-
-// Generated Alternative ClientToServer=
-//     TransactionData=(writes=ConstDict<ObjectFieldId, OneOf<None, Bytes>>, set_adds=ConstDict<IndexId, TupleOf<int64_t>>, set_removes=ConstDict<IndexId, TupleOf<int64_t>>, key_versions=TupleOf<ObjectFieldId>, index_versions=TupleOf<IndexId>, transaction_guid=int64_t)
-//     CompleteTransaction=(as_of_version=int64_t, transaction_guid=int64_t)
-//     Heartbeat=()
-//     DefineSchema=(name=String, definition=ConstDict<String, TypeDefinition>)
-//     LoadLazyObject=(schema=String, typename0=String, identity=int64_t)
-//     Subscribe=(schema=String, typename0=OneOf<None, String>, fieldname_and_value=OneOf<None, Anon39611216>, isLazy=bool)
-//     Flush=(guid=int64_t)
-//     Authenticate=(token=String)
-
-class ClientToServer_TransactionData;
-class ClientToServer_CompleteTransaction;
-class ClientToServer_Heartbeat;
-class ClientToServer_DefineSchema;
-class ClientToServer_LoadLazyObject;
-class ClientToServer_Subscribe;
-class ClientToServer_Flush;
-class ClientToServer_Authenticate;
-
-class ClientToServer {
-public:
-    enum class kind { TransactionData=0, CompleteTransaction=1, Heartbeat=2, DefineSchema=3, LoadLazyObject=4, Subscribe=5, Flush=6, Authenticate=7 };
-
-    static NamedTuple* TransactionData_Type;
-    static NamedTuple* CompleteTransaction_Type;
-    static NamedTuple* Heartbeat_Type;
-    static NamedTuple* DefineSchema_Type;
-    static NamedTuple* LoadLazyObject_Type;
-    static NamedTuple* Subscribe_Type;
-    static NamedTuple* Flush_Type;
-    static NamedTuple* Authenticate_Type;
-
-    static Alternative* getType();
-
-    static ClientToServer fromPython(PyObject* p) {
-        Alternative::layout* l = nullptr;
-        PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
-        return ClientToServer(l);
-    }
-
-    PyObject* toPython() {
-        return PyInstance::extractPythonObject((instance_ptr)&mLayout, getType());
-    }
-
-    ~ClientToServer() { getType()->destroy((instance_ptr)&mLayout); }
-    ClientToServer():mLayout(0) { getType()->constructor((instance_ptr)&mLayout); }
-    ClientToServer(kind k):mLayout(0) { ConcreteAlternative::Make(getType(), (int64_t)k)->constructor((instance_ptr)&mLayout); }
-    ClientToServer(const ClientToServer& in) { getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&in.mLayout); }
-    ClientToServer& operator=(const ClientToServer& other) { getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout); return *this; }
-
-    static ClientToServer TransactionData(const ConstDict<ObjectFieldId, OneOf<None, Bytes>>& writes, const ConstDict<IndexId, TupleOf<int64_t>>& set_adds, const ConstDict<IndexId, TupleOf<int64_t>>& set_removes, const TupleOf<ObjectFieldId>& key_versions, const TupleOf<IndexId>& index_versions, const int64_t& transaction_guid);
-    static ClientToServer CompleteTransaction(const int64_t& as_of_version, const int64_t& transaction_guid);
-    static ClientToServer Heartbeat();
-    static ClientToServer DefineSchema(const String& name, const ConstDict<String, TypeDefinition>& definition);
-    static ClientToServer LoadLazyObject(const String& schema, const String& typename0, const int64_t& identity);
-    static ClientToServer Subscribe(const String& schema, const OneOf<None, String>& typename0, const OneOf<None, Anon39611216>& fieldname_and_value, const bool& isLazy);
-    static ClientToServer Flush(const int64_t& guid);
-    static ClientToServer Authenticate(const String& token);
-
-    kind which() const { return (kind)mLayout->which; }
-
-    template <class F>
-    auto check(const F& f) {
-        if (isTransactionData()) { return f(*(ClientToServer_TransactionData*)this); }
-        if (isCompleteTransaction()) { return f(*(ClientToServer_CompleteTransaction*)this); }
-        if (isHeartbeat()) { return f(*(ClientToServer_Heartbeat*)this); }
-        if (isDefineSchema()) { return f(*(ClientToServer_DefineSchema*)this); }
-        if (isLoadLazyObject()) { return f(*(ClientToServer_LoadLazyObject*)this); }
-        if (isSubscribe()) { return f(*(ClientToServer_Subscribe*)this); }
-        if (isFlush()) { return f(*(ClientToServer_Flush*)this); }
-        if (isAuthenticate()) { return f(*(ClientToServer_Authenticate*)this); }
-    }
-
-    bool isTransactionData() const { return which() == kind::TransactionData; }
-    bool isCompleteTransaction() const { return which() == kind::CompleteTransaction; }
-    bool isHeartbeat() const { return which() == kind::Heartbeat; }
-    bool isDefineSchema() const { return which() == kind::DefineSchema; }
-    bool isLoadLazyObject() const { return which() == kind::LoadLazyObject; }
-    bool isSubscribe() const { return which() == kind::Subscribe; }
-    bool isFlush() const { return which() == kind::Flush; }
-    bool isAuthenticate() const { return which() == kind::Authenticate; }
-
-    // Accessors for members
-    ConstDict<ObjectFieldId, OneOf<None, Bytes>> writes() const;
-    ConstDict<IndexId, TupleOf<int64_t>> set_adds() const;
-    ConstDict<IndexId, TupleOf<int64_t>> set_removes() const;
-    TupleOf<ObjectFieldId> key_versions() const;
-    TupleOf<IndexId> index_versions() const;
-    int64_t transaction_guid() const;
-    int64_t as_of_version() const;
-    String name() const;
-    ConstDict<String, TypeDefinition> definition() const;
-    String schema() const;
-    OneOf<OneOf<None, String>,String> typename0() const;
-    int64_t identity() const;
-    OneOf<None, Anon39611216> fieldname_and_value() const;
-    bool isLazy() const;
-    int64_t guid() const;
-    String token() const;
-
-    Alternative::layout* getLayout() const { return mLayout; }
-protected:
-    explicit ClientToServer(Alternative::layout* l): mLayout(l) {}
-    Alternative::layout *mLayout;
-};
-
-template <>
-class TypeDetails<ClientToServer*> {
-public:
-    static Forward* getType() {
-        static Forward* t = new Forward(0, "ClientToServer");
-        return t;
-    }
-    static const uint64_t bytecount = sizeof(void*);
-};
-
-NamedTuple* ClientToServer::TransactionData_Type = NamedTuple::Make(
-    {TypeDetails<ConstDict<ObjectFieldId, OneOf<None, Bytes>>>::getType(), TypeDetails<ConstDict<IndexId, TupleOf<int64_t>>>::getType(), TypeDetails<ConstDict<IndexId, TupleOf<int64_t>>>::getType(), TypeDetails<TupleOf<ObjectFieldId>>::getType(), TypeDetails<TupleOf<IndexId>>::getType(), TypeDetails<int64_t>::getType()},
-    {"writes", "set_adds", "set_removes", "key_versions", "index_versions", "transaction_guid"}
-);
-
-NamedTuple* ClientToServer::CompleteTransaction_Type = NamedTuple::Make(
-    {TypeDetails<int64_t>::getType(), TypeDetails<int64_t>::getType()},
-    {"as_of_version", "transaction_guid"}
-);
-
-NamedTuple* ClientToServer::Heartbeat_Type = NamedTuple::Make(
-    {},
-    {}
-);
-
-NamedTuple* ClientToServer::DefineSchema_Type = NamedTuple::Make(
-    {TypeDetails<String>::getType(), TypeDetails<ConstDict<String, TypeDefinition>>::getType()},
-    {"name", "definition"}
-);
-
-NamedTuple* ClientToServer::LoadLazyObject_Type = NamedTuple::Make(
-    {TypeDetails<String>::getType(), TypeDetails<String>::getType(), TypeDetails<int64_t>::getType()},
-    {"schema", "typename0", "identity"}
-);
-
-NamedTuple* ClientToServer::Subscribe_Type = NamedTuple::Make(
-    {TypeDetails<String>::getType(), TypeDetails<OneOf<None, String>>::getType(), TypeDetails<OneOf<None, Anon39611216>>::getType(), TypeDetails<bool>::getType()},
-    {"schema", "typename0", "fieldname_and_value", "isLazy"}
-);
-
-NamedTuple* ClientToServer::Flush_Type = NamedTuple::Make(
-    {TypeDetails<int64_t>::getType()},
-    {"guid"}
-);
-
-NamedTuple* ClientToServer::Authenticate_Type = NamedTuple::Make(
-    {TypeDetails<String>::getType()},
-    {"token"}
-);
-
-// static
-Alternative* ClientToServer::getType() {
-    static Alternative* t = Alternative::Make("ClientToServer", {
-        {"TransactionData", TransactionData_Type},
-        {"CompleteTransaction", CompleteTransaction_Type},
-        {"Heartbeat", Heartbeat_Type},
-        {"DefineSchema", DefineSchema_Type},
-        {"LoadLazyObject", LoadLazyObject_Type},
-        {"Subscribe", Subscribe_Type},
-        {"Flush", Flush_Type},
-        {"Authenticate", Authenticate_Type}
-    }, {});
-    static bool once = false;
-    if (!once) {
-        once = true;
-        TypeDetails<ClientToServer*>::getType()->setTarget(t);
-        t = (Alternative*)t->guaranteeForwardsResolved([](void* p) { return (Type*)0; });
-    }
-    return t;
-}
-
-template <>
-class TypeDetails<ClientToServer> {
-public:
-    static Type* getType() {
-        static Type* t = ClientToServer::getType();
-        if (t->bytecount() != bytecount) {
-            throw std::runtime_error("ClientToServer somehow we have the wrong bytecount!");
-        }
-        return t;
-    }
-    static const uint64_t bytecount = sizeof(void*);
-};
-
-class ClientToServer_TransactionData : public ClientToServer {
-public:
-    static ConcreteAlternative* getType() {
-        static ConcreteAlternative* t = ConcreteAlternative::Make(ClientToServer::getType(), static_cast<int>(kind::TransactionData));
-        return t;
-    }
-    static Alternative* getAlternative() { return ClientToServer::getType(); }
-
-    ClientToServer_TransactionData():ClientToServer(kind::TransactionData) {}
-    ClientToServer_TransactionData( const ConstDict<ObjectFieldId, OneOf<None, Bytes>>& writes1,  const ConstDict<IndexId, TupleOf<int64_t>>& set_adds1,  const ConstDict<IndexId, TupleOf<int64_t>>& set_removes1,  const TupleOf<ObjectFieldId>& key_versions1,  const TupleOf<IndexId>& index_versions1,  const int64_t& transaction_guid1):ClientToServer(kind::TransactionData) {
-        writes() = writes1;
-        set_adds() = set_adds1;
-        set_removes() = set_removes1;
-        key_versions() = key_versions1;
-        index_versions() = index_versions1;
-        transaction_guid() = transaction_guid1;
-    }
-    ClientToServer_TransactionData(const ClientToServer_TransactionData& other):ClientToServer(kind::TransactionData) {
-        getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-    }
-    ClientToServer_TransactionData& operator=(const ClientToServer_TransactionData& other) {
-         getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-         return *this;
-    }
-    ~ClientToServer_TransactionData() {}
-
-    ConstDict<ObjectFieldId, OneOf<None, Bytes>>& writes() const { return *(ConstDict<ObjectFieldId, OneOf<None, Bytes>>*)(mLayout->data); }
-    ConstDict<IndexId, TupleOf<int64_t>>& set_adds() const { return *(ConstDict<IndexId, TupleOf<int64_t>>*)(mLayout->data + size1); }
-    ConstDict<IndexId, TupleOf<int64_t>>& set_removes() const { return *(ConstDict<IndexId, TupleOf<int64_t>>*)(mLayout->data + size1 + size2); }
-    TupleOf<ObjectFieldId>& key_versions() const { return *(TupleOf<ObjectFieldId>*)(mLayout->data + size1 + size2 + size3); }
-    TupleOf<IndexId>& index_versions() const { return *(TupleOf<IndexId>*)(mLayout->data + size1 + size2 + size3 + size4); }
-    int64_t& transaction_guid() const { return *(int64_t*)(mLayout->data + size1 + size2 + size3 + size4 + size5); }
-private:
-    static const int size1 = sizeof(ConstDict<ObjectFieldId, OneOf<None, Bytes>>);
-    static const int size2 = sizeof(ConstDict<IndexId, TupleOf<int64_t>>);
-    static const int size3 = sizeof(ConstDict<IndexId, TupleOf<int64_t>>);
-    static const int size4 = sizeof(TupleOf<ObjectFieldId>);
-    static const int size5 = sizeof(TupleOf<IndexId>);
-};
-
-ClientToServer ClientToServer::TransactionData(const ConstDict<ObjectFieldId, OneOf<None, Bytes>>& writes, const ConstDict<IndexId, TupleOf<int64_t>>& set_adds, const ConstDict<IndexId, TupleOf<int64_t>>& set_removes, const TupleOf<ObjectFieldId>& key_versions, const TupleOf<IndexId>& index_versions, const int64_t& transaction_guid) {
-    return ClientToServer_TransactionData(writes, set_adds, set_removes, key_versions, index_versions, transaction_guid);
-}
-
-class ClientToServer_CompleteTransaction : public ClientToServer {
-public:
-    static ConcreteAlternative* getType() {
-        static ConcreteAlternative* t = ConcreteAlternative::Make(ClientToServer::getType(), static_cast<int>(kind::CompleteTransaction));
-        return t;
-    }
-    static Alternative* getAlternative() { return ClientToServer::getType(); }
-
-    ClientToServer_CompleteTransaction():ClientToServer(kind::CompleteTransaction) {}
-    ClientToServer_CompleteTransaction( const int64_t& as_of_version1,  const int64_t& transaction_guid1):ClientToServer(kind::CompleteTransaction) {
-        as_of_version() = as_of_version1;
-        transaction_guid() = transaction_guid1;
-    }
-    ClientToServer_CompleteTransaction(const ClientToServer_CompleteTransaction& other):ClientToServer(kind::CompleteTransaction) {
-        getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-    }
-    ClientToServer_CompleteTransaction& operator=(const ClientToServer_CompleteTransaction& other) {
-         getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-         return *this;
-    }
-    ~ClientToServer_CompleteTransaction() {}
-
-    int64_t& as_of_version() const { return *(int64_t*)(mLayout->data); }
-    int64_t& transaction_guid() const { return *(int64_t*)(mLayout->data + size1); }
-private:
-    static const int size1 = sizeof(int64_t);
-};
-
-ClientToServer ClientToServer::CompleteTransaction(const int64_t& as_of_version, const int64_t& transaction_guid) {
-    return ClientToServer_CompleteTransaction(as_of_version, transaction_guid);
-}
-
-class ClientToServer_Heartbeat : public ClientToServer {
-public:
-    static ConcreteAlternative* getType() {
-        static ConcreteAlternative* t = ConcreteAlternative::Make(ClientToServer::getType(), static_cast<int>(kind::Heartbeat));
-        return t;
-    }
-    static Alternative* getAlternative() { return ClientToServer::getType(); }
-
-    ClientToServer_Heartbeat():ClientToServer(kind::Heartbeat) {}
-    ClientToServer_Heartbeat(const ClientToServer_Heartbeat& other):ClientToServer(kind::Heartbeat) {
-        getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-    }
-    ClientToServer_Heartbeat& operator=(const ClientToServer_Heartbeat& other) {
-         getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-         return *this;
-    }
-    ~ClientToServer_Heartbeat() {}
-
-private:
-};
-
-ClientToServer ClientToServer::Heartbeat() {
-    return ClientToServer_Heartbeat();
-}
-
-class ClientToServer_DefineSchema : public ClientToServer {
-public:
-    static ConcreteAlternative* getType() {
-        static ConcreteAlternative* t = ConcreteAlternative::Make(ClientToServer::getType(), static_cast<int>(kind::DefineSchema));
-        return t;
-    }
-    static Alternative* getAlternative() { return ClientToServer::getType(); }
-
-    ClientToServer_DefineSchema():ClientToServer(kind::DefineSchema) {}
-    ClientToServer_DefineSchema( const String& name1,  const ConstDict<String, TypeDefinition>& definition1):ClientToServer(kind::DefineSchema) {
-        name() = name1;
-        definition() = definition1;
-    }
-    ClientToServer_DefineSchema(const ClientToServer_DefineSchema& other):ClientToServer(kind::DefineSchema) {
-        getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-    }
-    ClientToServer_DefineSchema& operator=(const ClientToServer_DefineSchema& other) {
-         getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-         return *this;
-    }
-    ~ClientToServer_DefineSchema() {}
-
-    String& name() const { return *(String*)(mLayout->data); }
-    ConstDict<String, TypeDefinition>& definition() const { return *(ConstDict<String, TypeDefinition>*)(mLayout->data + size1); }
-private:
-    static const int size1 = sizeof(String);
-};
-
-ClientToServer ClientToServer::DefineSchema(const String& name, const ConstDict<String, TypeDefinition>& definition) {
-    return ClientToServer_DefineSchema(name, definition);
-}
-
-class ClientToServer_LoadLazyObject : public ClientToServer {
-public:
-    static ConcreteAlternative* getType() {
-        static ConcreteAlternative* t = ConcreteAlternative::Make(ClientToServer::getType(), static_cast<int>(kind::LoadLazyObject));
-        return t;
-    }
-    static Alternative* getAlternative() { return ClientToServer::getType(); }
-
-    ClientToServer_LoadLazyObject():ClientToServer(kind::LoadLazyObject) {}
-    ClientToServer_LoadLazyObject( const String& schema1,  const String& typename01,  const int64_t& identity1):ClientToServer(kind::LoadLazyObject) {
-        schema() = schema1;
-        typename0() = typename01;
-        identity() = identity1;
-    }
-    ClientToServer_LoadLazyObject(const ClientToServer_LoadLazyObject& other):ClientToServer(kind::LoadLazyObject) {
-        getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-    }
-    ClientToServer_LoadLazyObject& operator=(const ClientToServer_LoadLazyObject& other) {
-         getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-         return *this;
-    }
-    ~ClientToServer_LoadLazyObject() {}
-
-    String& schema() const { return *(String*)(mLayout->data); }
-    String& typename0() const { return *(String*)(mLayout->data + size1); }
-    int64_t& identity() const { return *(int64_t*)(mLayout->data + size1 + size2); }
-private:
-    static const int size1 = sizeof(String);
-    static const int size2 = sizeof(String);
-};
-
-ClientToServer ClientToServer::LoadLazyObject(const String& schema, const String& typename0, const int64_t& identity) {
-    return ClientToServer_LoadLazyObject(schema, typename0, identity);
-}
-
-class ClientToServer_Subscribe : public ClientToServer {
-public:
-    static ConcreteAlternative* getType() {
-        static ConcreteAlternative* t = ConcreteAlternative::Make(ClientToServer::getType(), static_cast<int>(kind::Subscribe));
-        return t;
-    }
-    static Alternative* getAlternative() { return ClientToServer::getType(); }
-
-    ClientToServer_Subscribe():ClientToServer(kind::Subscribe) {}
-    ClientToServer_Subscribe( const String& schema1,  const OneOf<None, String>& typename01,  const OneOf<None, Anon39611216>& fieldname_and_value1,  const bool& isLazy1):ClientToServer(kind::Subscribe) {
-        schema() = schema1;
-        typename0() = typename01;
-        fieldname_and_value() = fieldname_and_value1;
-        isLazy() = isLazy1;
-    }
-    ClientToServer_Subscribe(const ClientToServer_Subscribe& other):ClientToServer(kind::Subscribe) {
-        getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-    }
-    ClientToServer_Subscribe& operator=(const ClientToServer_Subscribe& other) {
-         getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-         return *this;
-    }
-    ~ClientToServer_Subscribe() {}
-
-    String& schema() const { return *(String*)(mLayout->data); }
-    OneOf<None, String>& typename0() const { return *(OneOf<None, String>*)(mLayout->data + size1); }
-    OneOf<None, Anon39611216>& fieldname_and_value() const { return *(OneOf<None, Anon39611216>*)(mLayout->data + size1 + size2); }
-    bool& isLazy() const { return *(bool*)(mLayout->data + size1 + size2 + size3); }
-private:
-    static const int size1 = sizeof(String);
-    static const int size2 = sizeof(OneOf<None, String>);
-    static const int size3 = sizeof(OneOf<None, Anon39611216>);
-};
-
-ClientToServer ClientToServer::Subscribe(const String& schema, const OneOf<None, String>& typename0, const OneOf<None, Anon39611216>& fieldname_and_value, const bool& isLazy) {
-    return ClientToServer_Subscribe(schema, typename0, fieldname_and_value, isLazy);
-}
-
-class ClientToServer_Flush : public ClientToServer {
-public:
-    static ConcreteAlternative* getType() {
-        static ConcreteAlternative* t = ConcreteAlternative::Make(ClientToServer::getType(), static_cast<int>(kind::Flush));
-        return t;
-    }
-    static Alternative* getAlternative() { return ClientToServer::getType(); }
-
-    ClientToServer_Flush():ClientToServer(kind::Flush) {}
-    ClientToServer_Flush( const int64_t& guid1):ClientToServer(kind::Flush) {
-        guid() = guid1;
-    }
-    ClientToServer_Flush(const ClientToServer_Flush& other):ClientToServer(kind::Flush) {
-        getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-    }
-    ClientToServer_Flush& operator=(const ClientToServer_Flush& other) {
-         getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-         return *this;
-    }
-    ~ClientToServer_Flush() {}
-
-    int64_t& guid() const { return *(int64_t*)(mLayout->data); }
-private:
-};
-
-ClientToServer ClientToServer::Flush(const int64_t& guid) {
-    return ClientToServer_Flush(guid);
-}
-
-class ClientToServer_Authenticate : public ClientToServer {
-public:
-    static ConcreteAlternative* getType() {
-        static ConcreteAlternative* t = ConcreteAlternative::Make(ClientToServer::getType(), static_cast<int>(kind::Authenticate));
-        return t;
-    }
-    static Alternative* getAlternative() { return ClientToServer::getType(); }
-
-    ClientToServer_Authenticate():ClientToServer(kind::Authenticate) {}
-    ClientToServer_Authenticate( const String& token1):ClientToServer(kind::Authenticate) {
-        token() = token1;
-    }
-    ClientToServer_Authenticate(const ClientToServer_Authenticate& other):ClientToServer(kind::Authenticate) {
-        getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-    }
-    ClientToServer_Authenticate& operator=(const ClientToServer_Authenticate& other) {
-         getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-         return *this;
-    }
-    ~ClientToServer_Authenticate() {}
-
-    String& token() const { return *(String*)(mLayout->data); }
-private:
-};
-
-ClientToServer ClientToServer::Authenticate(const String& token) {
-    return ClientToServer_Authenticate(token);
-}
-
-ConstDict<ObjectFieldId, OneOf<None, Bytes>> ClientToServer::writes() const {
-    if (isTransactionData())
-        return ((ClientToServer_TransactionData*)this)->writes();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"writes\"");
-}
-
-ConstDict<IndexId, TupleOf<int64_t>> ClientToServer::set_adds() const {
-    if (isTransactionData())
-        return ((ClientToServer_TransactionData*)this)->set_adds();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"set_adds\"");
-}
-
-ConstDict<IndexId, TupleOf<int64_t>> ClientToServer::set_removes() const {
-    if (isTransactionData())
-        return ((ClientToServer_TransactionData*)this)->set_removes();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"set_removes\"");
-}
-
-TupleOf<ObjectFieldId> ClientToServer::key_versions() const {
-    if (isTransactionData())
-        return ((ClientToServer_TransactionData*)this)->key_versions();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"key_versions\"");
-}
-
-TupleOf<IndexId> ClientToServer::index_versions() const {
-    if (isTransactionData())
-        return ((ClientToServer_TransactionData*)this)->index_versions();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"index_versions\"");
-}
-
-int64_t ClientToServer::transaction_guid() const {
-    if (isTransactionData())
-        return ((ClientToServer_TransactionData*)this)->transaction_guid();
-    if (isCompleteTransaction())
-        return ((ClientToServer_CompleteTransaction*)this)->transaction_guid();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"transaction_guid\"");
-}
-
-int64_t ClientToServer::as_of_version() const {
-    if (isCompleteTransaction())
-        return ((ClientToServer_CompleteTransaction*)this)->as_of_version();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"as_of_version\"");
-}
-
-String ClientToServer::name() const {
-    if (isDefineSchema())
-        return ((ClientToServer_DefineSchema*)this)->name();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"name\"");
-}
-
-ConstDict<String, TypeDefinition> ClientToServer::definition() const {
-    if (isDefineSchema())
-        return ((ClientToServer_DefineSchema*)this)->definition();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"definition\"");
-}
-
-String ClientToServer::schema() const {
-    if (isLoadLazyObject())
-        return ((ClientToServer_LoadLazyObject*)this)->schema();
-    if (isSubscribe())
-        return ((ClientToServer_Subscribe*)this)->schema();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"schema\"");
-}
-
-OneOf<OneOf<None, String>,String> ClientToServer::typename0() const {
-    if (isLoadLazyObject())
-        return OneOf<OneOf<None, String>,String>(((ClientToServer_LoadLazyObject*)this)->typename0());
-    if (isSubscribe())
-        return OneOf<OneOf<None, String>,String>(((ClientToServer_Subscribe*)this)->typename0());
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"typename0\"");
-}
-
-int64_t ClientToServer::identity() const {
-    if (isLoadLazyObject())
-        return ((ClientToServer_LoadLazyObject*)this)->identity();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"identity\"");
-}
-
-OneOf<None, Anon39611216> ClientToServer::fieldname_and_value() const {
-    if (isSubscribe())
-        return ((ClientToServer_Subscribe*)this)->fieldname_and_value();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"fieldname_and_value\"");
-}
-
-bool ClientToServer::isLazy() const {
-    if (isSubscribe())
-        return ((ClientToServer_Subscribe*)this)->isLazy();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"isLazy\"");
-}
-
-int64_t ClientToServer::guid() const {
-    if (isFlush())
-        return ((ClientToServer_Flush*)this)->guid();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"guid\"");
-}
-
-String ClientToServer::token() const {
-    if (isAuthenticate())
-        return ((ClientToServer_Authenticate*)this)->token();
-    throw std::runtime_error("\"ClientToServer\" subtype does not contain \"token\"");
-}
-
-// END Generated Alternative ClientToServer
+// END Generated NamedTuple Anon35597888
 
 // Generated Alternative A=
 //     Sub1=(b=int64_t, c=int64_t)
@@ -1430,8 +319,15 @@ public:
     static NamedTuple* Sub1_Type;
     static NamedTuple* Sub2_Type;
 
-    static Alternative* getType();
-
+    static Alternative* getType() {
+        PyObject* resolver = getOrSetTypeResolver();
+        if (!resolver)
+            throw std::runtime_error("{name}: no resolver");
+        PyObject* res = PyObject_CallMethod(resolver, "resolveTypeByName", "s", "A");
+        if (!res)
+            throw std::runtime_error("A: did not resolve");
+        return (Alternative*)PyInstance::unwrapTypeArgToTypePtr(res);
+    }
     static A fromPython(PyObject* p) {
         Alternative::layout* l = nullptr;
         PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
@@ -1474,16 +370,6 @@ protected:
     Alternative::layout *mLayout;
 };
 
-template <>
-class TypeDetails<A*> {
-public:
-    static Forward* getType() {
-        static Forward* t = new Forward(0, "A");
-        return t;
-    }
-    static const uint64_t bytecount = sizeof(void*);
-};
-
 NamedTuple* A::Sub1_Type = NamedTuple::Make(
     {TypeDetails<int64_t>::getType(), TypeDetails<int64_t>::getType()},
     {"b", "c"}
@@ -1493,21 +379,6 @@ NamedTuple* A::Sub2_Type = NamedTuple::Make(
     {TypeDetails<String>::getType(), TypeDetails<String>::getType()},
     {"d", "e"}
 );
-
-// static
-Alternative* A::getType() {
-    static Alternative* t = Alternative::Make("A", {
-        {"Sub1", Sub1_Type},
-        {"Sub2", Sub2_Type}
-    }, {});
-    static bool once = false;
-    if (!once) {
-        once = true;
-        TypeDetails<A*>::getType()->setTarget(t);
-        t = (Alternative*)t->guaranteeForwardsResolved([](void* p) { return (Type*)0; });
-    }
-    return t;
-}
 
 template <>
 class TypeDetails<A> {
@@ -1629,8 +500,15 @@ public:
     static NamedTuple* Sub2_Type;
     static NamedTuple* Sub3_Type;
 
-    static Alternative* getType();
-
+    static Alternative* getType() {
+        PyObject* resolver = getOrSetTypeResolver();
+        if (!resolver)
+            throw std::runtime_error("{name}: no resolver");
+        PyObject* res = PyObject_CallMethod(resolver, "resolveTypeByName", "s", "Overlap");
+        if (!res)
+            throw std::runtime_error("Overlap: did not resolve");
+        return (Alternative*)PyInstance::unwrapTypeArgToTypePtr(res);
+    }
     static Overlap fromPython(PyObject* p) {
         Alternative::layout* l = nullptr;
         PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
@@ -1665,23 +543,13 @@ public:
     bool isSub3() const { return which() == kind::Sub3; }
 
     // Accessors for members
-    OneOf<int64_t,bool,String> b() const;
-    OneOf<int64_t,TupleOf<String>> c() const;
+    OneOf<String,int64_t,bool> b() const;
+    OneOf<TupleOf<String>,int64_t> c() const;
 
     Alternative::layout* getLayout() const { return mLayout; }
 protected:
     explicit Overlap(Alternative::layout* l): mLayout(l) {}
     Alternative::layout *mLayout;
-};
-
-template <>
-class TypeDetails<Overlap*> {
-public:
-    static Forward* getType() {
-        static Forward* t = new Forward(0, "Overlap");
-        return t;
-    }
-    static const uint64_t bytecount = sizeof(void*);
 };
 
 NamedTuple* Overlap::Sub1_Type = NamedTuple::Make(
@@ -1698,22 +566,6 @@ NamedTuple* Overlap::Sub3_Type = NamedTuple::Make(
     {TypeDetails<int64_t>::getType()},
     {"b"}
 );
-
-// static
-Alternative* Overlap::getType() {
-    static Alternative* t = Alternative::Make("Overlap", {
-        {"Sub1", Sub1_Type},
-        {"Sub2", Sub2_Type},
-        {"Sub3", Sub3_Type}
-    }, {});
-    static bool once = false;
-    if (!once) {
-        once = true;
-        TypeDetails<Overlap*>::getType()->setTarget(t);
-        t = (Alternative*)t->guaranteeForwardsResolved([](void* p) { return (Type*)0; });
-    }
-    return t;
-}
 
 template <>
 class TypeDetails<Overlap> {
@@ -1821,269 +673,25 @@ Overlap Overlap::Sub3(const int64_t& b) {
     return Overlap_Sub3(b);
 }
 
-OneOf<int64_t,bool,String> Overlap::b() const {
+OneOf<String,int64_t,bool> Overlap::b() const {
     if (isSub1())
-        return OneOf<int64_t,bool,String>(((Overlap_Sub1*)this)->b());
+        return OneOf<String,int64_t,bool>(((Overlap_Sub1*)this)->b());
     if (isSub2())
-        return OneOf<int64_t,bool,String>(((Overlap_Sub2*)this)->b());
+        return OneOf<String,int64_t,bool>(((Overlap_Sub2*)this)->b());
     if (isSub3())
-        return OneOf<int64_t,bool,String>(((Overlap_Sub3*)this)->b());
+        return OneOf<String,int64_t,bool>(((Overlap_Sub3*)this)->b());
     throw std::runtime_error("\"Overlap\" subtype does not contain \"b\"");
 }
 
-OneOf<int64_t,TupleOf<String>> Overlap::c() const {
+OneOf<TupleOf<String>,int64_t> Overlap::c() const {
     if (isSub1())
-        return OneOf<int64_t,TupleOf<String>>(((Overlap_Sub1*)this)->c());
+        return OneOf<TupleOf<String>,int64_t>(((Overlap_Sub1*)this)->c());
     if (isSub2())
-        return OneOf<int64_t,TupleOf<String>>(((Overlap_Sub2*)this)->c());
+        return OneOf<TupleOf<String>,int64_t>(((Overlap_Sub2*)this)->c());
     throw std::runtime_error("\"Overlap\" subtype does not contain \"c\"");
 }
 
 // END Generated Alternative Overlap
-
-// Generated Alternative Bexpress=
-//     Leaf=(value=bool)
-//     BinOp=(left=Bexpress, op=String, right=Bexpress)
-//     UnaryOp=(op=String, right=Bexpress)
-
-class Bexpress_Leaf;
-class Bexpress_BinOp;
-class Bexpress_UnaryOp;
-
-class Bexpress {
-public:
-    enum class kind { Leaf=0, BinOp=1, UnaryOp=2 };
-
-    static NamedTuple* Leaf_Type;
-    static NamedTuple* BinOp_Type;
-    static NamedTuple* UnaryOp_Type;
-
-    static Alternative* getType();
-
-    static Bexpress fromPython(PyObject* p) {
-        Alternative::layout* l = nullptr;
-        PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
-        return Bexpress(l);
-    }
-
-    PyObject* toPython() {
-        return PyInstance::extractPythonObject((instance_ptr)&mLayout, getType());
-    }
-
-    ~Bexpress() { getType()->destroy((instance_ptr)&mLayout); }
-    Bexpress():mLayout(0) { getType()->constructor((instance_ptr)&mLayout); }
-    Bexpress(kind k):mLayout(0) { ConcreteAlternative::Make(getType(), (int64_t)k)->constructor((instance_ptr)&mLayout); }
-    Bexpress(const Bexpress& in) { getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&in.mLayout); }
-    Bexpress& operator=(const Bexpress& other) { getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout); return *this; }
-
-    static Bexpress Leaf(const bool& value);
-    static Bexpress BinOp(const Bexpress& left, const String& op, const Bexpress& right);
-    static Bexpress UnaryOp(const String& op, const Bexpress& right);
-
-    kind which() const { return (kind)mLayout->which; }
-
-    template <class F>
-    auto check(const F& f) {
-        if (isLeaf()) { return f(*(Bexpress_Leaf*)this); }
-        if (isBinOp()) { return f(*(Bexpress_BinOp*)this); }
-        if (isUnaryOp()) { return f(*(Bexpress_UnaryOp*)this); }
-    }
-
-    bool isLeaf() const { return which() == kind::Leaf; }
-    bool isBinOp() const { return which() == kind::BinOp; }
-    bool isUnaryOp() const { return which() == kind::UnaryOp; }
-
-    // Accessors for members
-    bool value() const;
-    Bexpress left() const;
-    String op() const;
-    Bexpress right() const;
-
-    Alternative::layout* getLayout() const { return mLayout; }
-protected:
-    explicit Bexpress(Alternative::layout* l): mLayout(l) {}
-    Alternative::layout *mLayout;
-};
-
-template <>
-class TypeDetails<Bexpress*> {
-public:
-    static Forward* getType() {
-        static Forward* t = new Forward(0, "Bexpress");
-        return t;
-    }
-    static const uint64_t bytecount = sizeof(void*);
-};
-
-NamedTuple* Bexpress::Leaf_Type = NamedTuple::Make(
-    {TypeDetails<bool>::getType()},
-    {"value"}
-);
-
-NamedTuple* Bexpress::BinOp_Type = NamedTuple::Make(
-    {TypeDetails<Bexpress*>::getType(), TypeDetails<String>::getType(), TypeDetails<Bexpress*>::getType()},
-    {"left", "op", "right"}
-);
-
-NamedTuple* Bexpress::UnaryOp_Type = NamedTuple::Make(
-    {TypeDetails<String>::getType(), TypeDetails<Bexpress*>::getType()},
-    {"op", "right"}
-);
-
-// static
-Alternative* Bexpress::getType() {
-    static Alternative* t = Alternative::Make("Bexpress", {
-        {"Leaf", Leaf_Type},
-        {"BinOp", BinOp_Type},
-        {"UnaryOp", UnaryOp_Type}
-    }, {});
-    static bool once = false;
-    if (!once) {
-        once = true;
-        TypeDetails<Bexpress*>::getType()->setTarget(t);
-        t = (Alternative*)t->guaranteeForwardsResolved([](void* p) { return (Type*)0; });
-    }
-    return t;
-}
-
-template <>
-class TypeDetails<Bexpress> {
-public:
-    static Type* getType() {
-        static Type* t = Bexpress::getType();
-        if (t->bytecount() != bytecount) {
-            throw std::runtime_error("Bexpress somehow we have the wrong bytecount!");
-        }
-        return t;
-    }
-    static const uint64_t bytecount = sizeof(void*);
-};
-
-class Bexpress_Leaf : public Bexpress {
-public:
-    static ConcreteAlternative* getType() {
-        static ConcreteAlternative* t = ConcreteAlternative::Make(Bexpress::getType(), static_cast<int>(kind::Leaf));
-        return t;
-    }
-    static Alternative* getAlternative() { return Bexpress::getType(); }
-
-    Bexpress_Leaf():Bexpress(kind::Leaf) {}
-    Bexpress_Leaf( const bool& value1):Bexpress(kind::Leaf) {
-        value() = value1;
-    }
-    Bexpress_Leaf(const Bexpress_Leaf& other):Bexpress(kind::Leaf) {
-        getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-    }
-    Bexpress_Leaf& operator=(const Bexpress_Leaf& other) {
-         getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-         return *this;
-    }
-    ~Bexpress_Leaf() {}
-
-    bool& value() const { return *(bool*)(mLayout->data); }
-private:
-};
-
-Bexpress Bexpress::Leaf(const bool& value) {
-    return Bexpress_Leaf(value);
-}
-
-class Bexpress_BinOp : public Bexpress {
-public:
-    static ConcreteAlternative* getType() {
-        static ConcreteAlternative* t = ConcreteAlternative::Make(Bexpress::getType(), static_cast<int>(kind::BinOp));
-        return t;
-    }
-    static Alternative* getAlternative() { return Bexpress::getType(); }
-
-    Bexpress_BinOp():Bexpress(kind::BinOp) {}
-    Bexpress_BinOp( const Bexpress& left1,  const String& op1,  const Bexpress& right1):Bexpress(kind::BinOp) {
-        left() = left1;
-        op() = op1;
-        right() = right1;
-    }
-    Bexpress_BinOp(const Bexpress_BinOp& other):Bexpress(kind::BinOp) {
-        getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-    }
-    Bexpress_BinOp& operator=(const Bexpress_BinOp& other) {
-         getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-         return *this;
-    }
-    ~Bexpress_BinOp() {}
-
-    Bexpress& left() const { return *(Bexpress*)(mLayout->data); }
-    String& op() const { return *(String*)(mLayout->data + size1); }
-    Bexpress& right() const { return *(Bexpress*)(mLayout->data + size1 + size2); }
-private:
-    static const int size1 = sizeof(Bexpress);
-    static const int size2 = sizeof(String);
-};
-
-Bexpress Bexpress::BinOp(const Bexpress& left, const String& op, const Bexpress& right) {
-    return Bexpress_BinOp(left, op, right);
-}
-
-class Bexpress_UnaryOp : public Bexpress {
-public:
-    static ConcreteAlternative* getType() {
-        static ConcreteAlternative* t = ConcreteAlternative::Make(Bexpress::getType(), static_cast<int>(kind::UnaryOp));
-        return t;
-    }
-    static Alternative* getAlternative() { return Bexpress::getType(); }
-
-    Bexpress_UnaryOp():Bexpress(kind::UnaryOp) {}
-    Bexpress_UnaryOp( const String& op1,  const Bexpress& right1):Bexpress(kind::UnaryOp) {
-        op() = op1;
-        right() = right1;
-    }
-    Bexpress_UnaryOp(const Bexpress_UnaryOp& other):Bexpress(kind::UnaryOp) {
-        getType()->copy_constructor((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-    }
-    Bexpress_UnaryOp& operator=(const Bexpress_UnaryOp& other) {
-         getType()->assign((instance_ptr)&mLayout, (instance_ptr)&other.mLayout);
-         return *this;
-    }
-    ~Bexpress_UnaryOp() {}
-
-    String& op() const { return *(String*)(mLayout->data); }
-    Bexpress& right() const { return *(Bexpress*)(mLayout->data + size1); }
-private:
-    static const int size1 = sizeof(String);
-};
-
-Bexpress Bexpress::UnaryOp(const String& op, const Bexpress& right) {
-    return Bexpress_UnaryOp(op, right);
-}
-
-bool Bexpress::value() const {
-    if (isLeaf())
-        return ((Bexpress_Leaf*)this)->value();
-    throw std::runtime_error("\"Bexpress\" subtype does not contain \"value\"");
-}
-
-Bexpress Bexpress::left() const {
-    if (isBinOp())
-        return ((Bexpress_BinOp*)this)->left();
-    throw std::runtime_error("\"Bexpress\" subtype does not contain \"left\"");
-}
-
-String Bexpress::op() const {
-    if (isBinOp())
-        return ((Bexpress_BinOp*)this)->op();
-    if (isUnaryOp())
-        return ((Bexpress_UnaryOp*)this)->op();
-    throw std::runtime_error("\"Bexpress\" subtype does not contain \"op\"");
-}
-
-Bexpress Bexpress::right() const {
-    if (isBinOp())
-        return ((Bexpress_BinOp*)this)->right();
-    if (isUnaryOp())
-        return ((Bexpress_UnaryOp*)this)->right();
-    throw std::runtime_error("\"Bexpress\" subtype does not contain \"right\"");
-}
-
-// END Generated Alternative Bexpress
 
 // Generated NamedTuple NamedTupleTwoStrings
 //    X=String
@@ -2310,110 +918,6 @@ public:
 };
 
 // END Generated NamedTuple NamedTupleBoolIntStr
-
-// Generated NamedTuple Choice
-//    A=NamedTupleTwoStrings
-//    B=Bexpress
-class Choice {
-public:
-    typedef NamedTupleTwoStrings A_type;
-    typedef Bexpress B_type;
-    A_type& A() const { return *(A_type*)(data); }
-    B_type& B() const { return *(B_type*)(data + size1); }
-    static NamedTuple* getType() {
-        static NamedTuple* t = NamedTuple::Make({
-                TypeDetails<Choice::A_type>::getType(),
-                TypeDetails<Choice::B_type>::getType()
-            },{
-                "A",
-                "B"
-            });
-        return t;
-        }
-
-    static Choice fromPython(PyObject* p) {
-        Choice l;
-        PyInstance::copyConstructFromPythonInstance(getType(), (instance_ptr)&l, p, true);
-        return l;
-    }
-
-    PyObject* toPython() {
-        return PyInstance::extractPythonObject((instance_ptr)this, getType());
-    }
-
-    Choice& operator = (const Choice& other) {
-        A() = other.A();
-        B() = other.B();
-        return *this;
-    }
-
-    Choice(const Choice& other) {
-        new (&A()) A_type(other.A());
-        new (&B()) B_type(other.B());
-    }
-
-    ~Choice() {
-        B().~B_type();
-        A().~A_type();
-    }
-
-    Choice() {
-        bool initA = false;
-        bool initB = false;
-        try {
-            new (&A()) A_type();
-            initA = true;
-            new (&B()) B_type();
-            initB = true;
-        } catch(...) {
-            try {
-                if (initB) B().~B_type();
-                if (initA) A().~A_type();
-            } catch(...) {
-            }
-            throw;
-        }
-    }
-
-    Choice(const A_type& A_val, const B_type& B_val) {
-        bool initA = false;
-        bool initB = false;
-        try {
-            new (&A()) A_type(A_val);
-            initA = true;
-            new (&B()) B_type(B_val);
-            initB = true;
-        } catch(...) {
-            try {
-                if (initB) B().~B_type();
-                if (initA) A().~A_type();
-            } catch(...) {
-            }
-            throw;
-        }
-    }
-private:
-    static const int size1 = sizeof(A_type);
-    static const int size2 = sizeof(B_type);
-    uint8_t data[size1 + size2];
-};
-
-template <>
-class TypeDetails<Choice> {
-public:
-    static Type* getType() {
-        static Type* t = Choice::getType();
-        if (t->bytecount() != bytecount) {
-            throw std::runtime_error("Choice somehow we have the wrong bytecount!");
-        }
-        return t;
-    }
-    static const uint64_t bytecount = 
-        sizeof(Choice::A_type) +
-        sizeof(Choice::B_type);
-};
-
-// END Generated NamedTuple Choice
 
 // Generated NamedTuple NamedTupleIntFloatDesc
 //    a=OneOf<int64_t, double, bool>
@@ -2746,16 +1250,16 @@ public:
 // END Generated NamedTuple NamedTupleAttrAndValues
 
 // Generated Tuple AnonTest
-//    a0=Dict<Anon41033936, String>
-//    a1=ConstDict<String, OneOf<bool, Anon41042080>>
-//    a2=ListOf<Anon41033936>
-//    a3=TupleOf<Anon41057728>
+//    a0=Dict<Anon35575456, String>
+//    a1=ConstDict<String, OneOf<bool, Anon35583280>>
+//    a2=ListOf<Anon35575456>
+//    a3=TupleOf<Anon35597888>
 class AnonTest {
 public:
-    typedef Dict<Anon41033936, String> a0_type;
-    typedef ConstDict<String, OneOf<bool, Anon41042080>> a1_type;
-    typedef ListOf<Anon41033936> a2_type;
-    typedef TupleOf<Anon41057728> a3_type;
+    typedef Dict<Anon35575456, String> a0_type;
+    typedef ConstDict<String, OneOf<bool, Anon35583280>> a1_type;
+    typedef ListOf<Anon35575456> a2_type;
+    typedef TupleOf<Anon35597888> a3_type;
     a0_type& a0() const { return *(a0_type*)(data); }
     a1_type& a1() const { return *(a1_type*)(data + size1); }
     a2_type& a2() const { return *(a2_type*)(data + size1 + size2); }
