+++ typed_python/direct_types/generate_named_tuple.py
@@ -15,16 +15,17 @@
 #   limitations under the License.
 
 
-def gen_named_tuple_type(name, **kwargs):
+def gen_named_tuple_type(full_name, **kwargs):
     """
     Generate direct c++ wrapper code for a particular NamedTuple type.
 
     Args:
-        name: string name of this NamedTuple type
+        full_name: fully specified dotted name from codebase, module.class.subclass. ... .typename
         **kwargs: keys are element names and values are corresponding python Types
     Returns:
         A list of strings, containing c++ code implementing this wrapper.
     """
+    name = full_name.rsplit('.', 1)[-1]
     items = kwargs.items()
     keys = kwargs.keys()
     revkeys = list(keys)[::-1]
