+++ typed_python/direct_types/generate_types.py
@@ -17,12 +17,13 @@
 import sys
 import argparse
 import traceback
+
+from typed_python.Codebase import Codebase
 from typed_python._types import Dict, ConstDict, NamedTuple, Tuple, ListOf, TupleOf, OneOf, Alternative
+from typed_python._types import getOrSetTypeResolver
 from typed_python.direct_types.generate_tuple import gen_tuple_type
 from typed_python.direct_types.generate_named_tuple import gen_named_tuple_type
 from typed_python.direct_types.generate_alternative import gen_alternative_type
-from object_database.schema import SchemaDefinition, ObjectFieldId, IndexId, FieldDefinition, TypeDefinition
-from object_database.messages import ClientToServer
 
 
 cpp_type_mapping = {}
@@ -55,8 +56,8 @@ def cpp_type(py_type) -> (str, dict):
 
     Returns:
         tuple (string, dict)
-        string is the c++ type name
-        dict (str->Type) of anonymous type prerequisites for this time
+            string is the c++ type name
+            dict (str->Type) of anonymous type prerequisites for this time
     """
     simple_cats = {
         'None': 'None',
@@ -87,11 +88,12 @@ def cpp_type(py_type) -> (str, dict):
         for e in result:
             prerequisites.update(e[1])
         cpp_name = 'OneOf<{}>'.format(', '.join([e[0] for e in result]))
-    elif cat == 'Forward':
-        cpp_name = str(py_type)[8:-2] + '*'  # just for now!
+    # Forwards are no longer supported here: types must be fully defined (resolved) before generating direct c++ types
+    # elif cat == 'Forward':
+    #    cpp_name = str(py_type)[8:-2] + '*'  # just for now!
     elif cat in ('Tuple', 'NamedTuple', 'Alternative'):
         if py_type in cpp_type_mapping:
-            cpp_name = cpp_type_mapping[py_type]
+            cpp_name = cpp_type_mapping[py_type].rsplit(".", 1)[-1]
         else:
             name = anon_name(py_type)
             prerequisites[name] = py_type
@@ -151,6 +153,94 @@ def typed_python_codegen(**kwargs):
     return typed_python_codegen(**anon_types) + ret
 
 
+def typed_python_codegen2(**kwargs):
+    """Generates direct c++ wrappers for Tuple, NamedTuple, and Alternative types
+
+    Args:
+        **kwargs: Dict of name -> Type
+    Returns:
+        A list of strings, containing c++ code implementing wrappers for these types.
+    """
+    ret = []
+    if len(kwargs.items()) == 0:
+        return ret
+    for k, v in kwargs.items():
+        k = avoid_cpp_keywords(k)
+        if v.__typed_python_category__ in ('Tuple', 'NamedTuple', 'Alternative'):
+            if v.__typed_python_category__ == 'Tuple':
+                ret += gen_tuple_type(k, *[cpp_type(t)[0] for t in v.ElementTypes])
+                # for t in v.ElementTypes:
+                #     anon_types.update(cpp_type(t)[1])
+            elif v.__typed_python_category__ == 'NamedTuple':
+                ret += gen_named_tuple_type(k, **{avoid_cpp_keywords(n): cpp_type(t)[0] for n, t in zip(v.ElementNames, v.ElementTypes)})
+                # for t in v.ElementTypes:
+                #     anon_types.update(cpp_type(t)[1])
+            elif v.__typed_python_category__ == 'Alternative':
+                d = {nt.Name: [(avoid_cpp_keywords(a), cpp_type(t)[0])
+                               for a, t in zip(nt.ElementType.ElementNames, nt.ElementType.ElementTypes)]
+                     for nt in v.__typed_python_alternatives__}
+                ret += gen_alternative_type(k, d)
+                # for nt in v.__typed_python_alternatives__:
+                #     for t in nt.ElementType.ElementTypes:
+                #         anon_types.update(cpp_type(t)[1])
+
+            cpp_type_mapping[v] = k
+    return ret
+
+
+def generate_cpp(codebase, t, cpp, verbose=False, produce_code=True):
+    if t in cpp:
+        return
+    cat = t.__typed_python_category__
+    if cat in ('Tuple', 'NamedTuple'):
+        for s in t.ElementTypes:
+            generate_cpp(codebase, s, cpp, verbose)
+    elif cat == 'Alternative':
+        for s in t.__typed_python_alternatives__:
+            generate_cpp(codebase, s.ElementType, cpp, verbose, False)
+    elif cat in ('TupleOf', 'ListOf'):
+        generate_cpp(codebase, t.ElementType, cpp, verbose)
+    elif cat == 'OneOf':
+        for s in t.Types:
+            generate_cpp(codebase, s, cpp, verbose)
+    elif cat in ('Dict', 'ConstDict'):
+        generate_cpp(codebase, t.KeyType, cpp, verbose)
+        generate_cpp(codebase, t.ValueType, cpp, verbose)
+
+    if produce_code and cat in ('Tuple', 'NamedTuple', 'Alternative'):
+        name = codebase.serializationContext.objToName.get(id(t), anon_name(t))
+        if verbose:
+            print(f'    Gen {t.__name__} = {name}')
+            if name in cpp:
+                print('        already defined')
+        cpp[name] = t
+
+
+def generate_from_codebase(codebase, dest, one=None, verbose=False):
+    cpp = {}
+    if one:
+        items = [(one, codebase.serializationContext.nameToObject[one])]
+    else:
+        items = codebase.serializationContext.nameToObject.items()
+    for n, t in items:
+        if hasattr(t, '__typed_python_category__'):
+            if verbose:
+                print(f'{n} {t.__typed_python_category__}')
+            if t.__typed_python_category__ in ('Tuple', 'NamedTuple', 'Alternative'):
+                generate_cpp(codebase, t, cpp, verbose)
+    with open(dest, 'w') as f:
+        f.write('#pragma once\n')
+        f.writelines(typed_python_codegen2(**cpp))
+
+
+class resolver:
+    def __init__(self, codebase):
+        self._codebase = codebase
+
+    def resolveTypeByName(self, s):
+        return self._codebase.serializationContext.nameToObject[s]
+
+
 def generate_some_types(dest):
     """Generates direct c++ type wrappers for testing.
 
@@ -161,7 +251,7 @@ def generate_some_types(dest):
     """
     Bexpress = lambda: Bexpress
     Bexpress = Alternative(
-        "BooleanExpr",
+        "Bexpress",
         Leaf={
             "value": bool
         },
@@ -173,24 +263,24 @@ def generate_some_types(dest):
         UnaryOp={
             "op": str,
             "right": Bexpress
-        }
-    )
+        })
+    # A = Alternative('A', Sub1={'b': int, 'c': int}, Sub2={'d': str, 'e': str})
+    # Overlap = Alternative('Overlap', Sub1={'b': bool, 'c': int}, Sub2={'b': str, 'c': TupleOf(str)}, Sub3={'b': int})
     with open(dest, 'w') as f:
-        f.write("#pragma once\n")
+        f.write('#pragma once\n')
         f.writelines(typed_python_codegen(
-            ObjectFieldId=ObjectFieldId,
-            IndexId=IndexId,
-            FieldDefinition=FieldDefinition,
-            TypeDefinition=TypeDefinition,
-            SchemaDefinition=SchemaDefinition,
-            ClientToServer=ClientToServer,
-            A=Alternative('A', Sub1={'b': int, 'c': int}, Sub2={'d': str, 'e': str}),
-            Overlap=Alternative('Overlap', Sub1={'b': bool, 'c': int}, Sub2={'b': str, 'c': TupleOf(str)},
-                                Sub3={'b': int}),
-            Bexpress=Bexpress,
+            # ObjectFieldId=ObjectFieldId,
+            # IndexId=IndexId,
+            # FieldDefinition=FieldDefinition,
+            # TypeDefinition=TypeDefinition,
+            # SchemaDefinition=SchemaDefinition,
+            # ClientToServer=ClientToServer,
+            # A=A,
+            # Overlap=Overlap,
+            # Bexpress=Bexpress,
             NamedTupleTwoStrings=NamedTuple(X=str, Y=str),
             NamedTupleBoolIntStr=NamedTuple(b=bool, i=int, s=str),
-            Choice=NamedTuple(A=NamedTuple(X=str, Y=str), B=Bexpress),
+            # Choice=NamedTuple(A=NamedTuple(X=str, Y=str), B=Bexpress),
             NamedTupleIntFloatDesc=NamedTuple(a=OneOf(int, float, bool), b=float, desc=str),
             NamedTupleBoolListOfInt=NamedTuple(X=bool, Y=ListOf(int)),
             NamedTupleAttrAndValues=NamedTuple(attributes=TupleOf(str), values=TupleOf(int)),
@@ -210,14 +300,28 @@ def main(argv):
     parser = argparse.ArgumentParser(description='Generate types')
     parser.add_argument('dest', nargs='?', default='DefaultGeneratedTestTypes.hpp')
     parser.add_argument('-t', '--testTypes', action='store_true')
+    parser.add_argument('-c', '--testTypes2', action='store_true')
+    parser.add_argument('-d', '--testTypes3', action='store_true')
+    parser.add_argument('-e', '--testTypes4', action='store_true')
+    parser.add_argument('-v', '--verbose', action='store_true')
     args = parser.parse_args()
 
-    if args.testTypes:
-        try:
+    try:
+        if args.testTypes:
             generate_some_types(args.dest)
-        except Exception:
-            print("FAILED:\n", traceback.format_exc())
-            return 1
+        elif args.testTypes2:
+            codebase = Codebase.FromRootlevelPath("object_database")
+            getOrSetTypeResolver(resolver(codebase))
+            generate_from_codebase(codebase, args.dest, "object_database.messages.ClientToServer", verbose=args.verbose)
+        elif args.testTypes3:
+            codebase = Codebase.FromRootlevelPath("typed_python")
+            generate_from_codebase(codebase, args.dest, "typed_python.python_ast.Expr", verbose=args.verbose)
+        elif args.testTypes4:
+            codebase = Codebase.FromRootlevelPath("object_database", verbose=args.verbose)
+            generate_from_codebase(codebase, args.dest)
+    except Exception:
+        print("FAILED:\n", traceback.format_exc())
+        return 1
     return 0
