+++ typed_python/util.cpp
