+++ typed_python/util.hpp
@@ -265,4 +265,3 @@ void iterate(PyObject* o, func_type f) {
         throw PythonExceptionSet();
     }
 }
-
